# Reduce TrieNode allocations