… src/Nethermind/Nethermind.Trie/TrieNode.cs