@Scooletz 2023-03-21
Maybe I can change the implementation to use `Rlp.ValueDecoderContext`?