@Scooletz 2023-03-21
Doing a full sync on teku mainnet. Once it's synced and working will confirm and merge if the results are consistent.