@Scooletz 2023-03-21
The teku mainnet has been synced with this and now is operational. Waiting for a few hours of running and will be merging if nothing contradicts the findings above.