@Scooletz 2023-03-22
I confirm that with the usage of the Rlp context reading the version behaves as it was with the copied `RlpStreamReader`, reducing the measured alloc ratio by at least 10%. Merging.